Split `common::points` into `common::query` and `common::update`
