+++ src/common/mod.rs
@@ -9,9 +9,11 @@ pub mod inference;
 pub mod metrics;
 pub mod points;
 pub mod pyroscope_state;
+pub mod query;
 pub mod snapshots;
 pub mod stacktrace;
 pub mod strings;
 pub mod telemetry;
 pub mod telemetry_ops;
 pub mod telemetry_reporting;
+pub mod update;
